Move coercion related methods to a util class
